Create BDD test

+++ src/main/java/app/controller/ConfigurationController.java
@@ -2,6 +2,7 @@ package app.controller;
 
 import app.entity.Configuration;
 import app.entity.wrapper.ConfigurationWrapper;
+import app.service.AccessTokenService;
 import app.service.ConfigurationService;
 import lombok.NonNull;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,16 +19,30 @@ public class ConfigurationController {
     @Autowired
     ConfigurationService service;
 
+    @Autowired
+    AccessTokenService accessTokenService;
+
     @GetMapping("/configuration")
-    public ResponseEntity<List<Configuration>> getAll() {
+    public ResponseEntity<List<Configuration>> getAll(@RequestAttribute String token) {
+        if (!accessTokenService.isTokenValid(token)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return new ResponseEntity<>(service.findAll(), HttpStatus.OK);
     }
+
     @PostMapping("/configuration")
-    public ResponseEntity save(@NonNull @RequestBody ConfigurationWrapper configurationWrapper) {
+    public ResponseEntity save(@NonNull @RequestBody ConfigurationWrapper configurationWrapper, @RequestAttribute String token) {
+        if (!accessTokenService.isTokenValid(token)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.status(service.saveFlow(configurationWrapper.getConfigurations())).build();
     }
+
     @DeleteMapping("/configuration/{configKey}")
-    public ResponseEntity delete(@PathVariable String configKey) {
+    public ResponseEntity delete(@PathVariable String configKey, @RequestAttribute String token) {
+        if (!accessTokenService.isTokenValid(token)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.status(service.deleteFlow(configKey)).build();
     }
 }
